+++ R/efficient.R
@@ -94,35 +94,37 @@ conditional_parms <- function(s, samples) {
 
 #' Test that the sampler has successfully adapted
 #'
-#' @param stage_samples The samples store for the currently running stage.
-#' @param pmwgs The full pmwgs object with all samples from previous runs.
+#' @param pmwgs The full pmwgs object with all samples
 #' @param n_unique The number of unique samples to look for in random effects
-#'   for each participant.
+#'   for each subject.
 #' @param i The number for the current iteration of the sampler
 #'
 #' @return A string representing successful/unsuccessful adaptation. Can be one
 #'   of c("success", "continue", "increase")
 #'
 #' @keywords internal
-test_sampler_adapted <- function(stage_samples, pmwgs, n_unique, i) {
-  if (check_adapted(stage_samples$alpha, unq_vals = n_unique)) {
+test_sampler_adapted <- function(pmwgs, n_unique, i) {
+  if (i < n_unique) {
+    return("continue")
+  }
+  test_samples <- extract_samples(pmwgs, stage = "adapt")
+  if (check_adapted(test_samples$alpha, unq_vals = n_unique)) {
     message("Enough unique values detected: ", n_unique)
     message("Testing proposal distribution creation")
     attempt <- try({
-      tmp_sampler <- update_sampler(pmwgs, stage_samples)
       lapply(
-        X = 1:tmp_sampler$n_subjects,
+        X = 1:pmwgs$n_subjects,
         FUN = conditional_parms,
-        samples = extract_samples(tmp_sampler)
+        samples = test_samples
       )
     })
     if (class(attempt) == "try-error") {
-      warning("An error was encountered creating proposal distribution")
-      warning("Increasing required unique values")
+      warning("An problem was encountered creating proposal distribution")
+      warning("Increasing required unique values and continuing adaptation")
       return("increase")
     }
     else {
-      message("Adapted after ", i, "iterations - stopping early")
+      message("Successfully adapted after ", i, "iterations - stopping early")
       return("success")
     }
   }
@@ -140,10 +142,10 @@ test_sampler_adapted <- function(stage_samples, pmwgs, n_unique, i) {
 check_adapted <- function(samples, unq_vals = 20) {
   # Only need to check uniqueness for one parameter
   first_par <- samples[1, , ]
-  all(
-    lapply(
-      apply(first_par, 1, unique),
-      length
-    ) > unq_vals
-  )
+  # Split the matrix into a list of vectors by subject
+  # Needed for the case where every sample is unique for all subjects
+  first_par_list <- split(first_par, seq(NROW(first_par)))
+  # Get unique pars (new accepted particles) and check length for
+  # all subjects is greater than unq_vals
+  all(lapply(lapply(first_par_list, unique), length) > unq_vals)
 }