+++ R/pmwg-package.R
@@ -1,12 +1,11 @@
 #' pmwg: Particle Metropolis Within Gibbs.
 #'
 #' The pmwg package provides a general purpose implementation of the
-#' sampling techniques outlined in
-#' \href{https://doi.org/10.1016/j.jmp.2020.102368}{Gunawan et al. (2020)}.
-#' The user of this package is required to provide their own log likelihood
-#' function, but given this the functions provided can estimate model
-#' parameters, the full covariance matrix and subject random effects in a
-#' hierarchical Bayesian way.
+#' sampling techniques outlined in Gunawan et al. (2020)
+#' \doi{10.1016/j.jmp.2020.102368}. The user of this package is required to
+#' provide their own log likelihood function, but given this the functions
+#' provided can estimate model parameters, the full covariance matrix and
+#' subject random effects in a hierarchical Bayesian way.
 #'
 #' @section Documentation:
 #' The documentation found at \url{https://newcastlecl.github.io/samplerDoc/}
@@ -17,7 +16,7 @@
 #' @section User input:
 #' The user is expected to provide a data source in a format that is compatible
 #' with R data.frame methods. This data must have at least one column named
-#' `subject` that has a unique identifier for each participants data.
+#' `subject` that has a unique identifier for each subject's data.
 #'
 #' Additionally the user should provide a function that when given a set of
 #' parameter estimates and the data for a single subject return the log of the